Split out most of mars.h into globals.h/globals.c
